Change how we check for categories
Categories are not reliable because a page gets
added to a category async via a job queue. This is
quite an issue in our case since we want to build
a reliable interface every time. This is a quick
and dirty solution for the problem, since we add
the categories automatically via CoureEditor it
should not be an issue (but it is bad code).

Signed-off-by: Cristian Baldi <bld.cris.96@gmail.com>