Style various kind of images
The style was added for the kinds of images which
are styleable, some kinds (frameless/base) don't
have any additional classes added to their html
structure so it is not possible to style them
(well, it is possible but it will ruin the display
of some templates which we can't/won't change
[Collection], will get on this after templating
work is done)

Known bugs: inserting a big enough image in Visual
Editor and setting the style to "Base" will glitch
out VE expanding the image to the right, even after the page ends.
No fix is  planned from me since it is a pain to work with VE
classes and it is bugged even on upstream. The problem is correctly handled
in display mode