Add some JS code to wrap environment content inside a div to allow styling.
Unfortunately this is the bad way of doing it, since the good way does not work. The good way was opening the
environment tag inside the BeginEnv template and closing it in the EndEnv template: this does not work as
expected because Visual Editor decide to join the content together as an "Inclusion"; Too bad this feature
breaks our intent.