Enable font selection on user's preferences
This commit add the ability to choose four different type of
fonts optimized for a better reading of a page. To enable or
change them a user should go to Special:Preferences and look
for Aspect->Skin Font.
If the font is not selected it doesn't get downloaded, in
order to keep the loading of the assets the minimum possible